After storing a product and price, return history
This might need some refinement for find/create operations, but this is
better than starting from scratch on the query since we already have a
product model to grab data for. I'm not sure if there is a better way of
converting product history per site to the way that I think it should
look.